Create copy protection for the site Fixes #14
Created a function that redirects or leaves the site
Includes a mailto: link to contact support (me) in case of errors
calling the function with a string argument allows the file to live
in a folder.

Further work is required to extend functionality of the "Beta" path
additionally the function should determine if its being passed properly
URIencoded paths. Will create an additional issue for post launch